Fixed library references to use KiCad 4 backport of KiCad 5 libraries
This project is currently implemented in KiCad 4.0.x, and will remain as
such until KiCad 5 is released and easily available.

However, there have been a lot of improvements to the component
libraries since 4.0.x was frozen. As such, this project was originally
developed with a haphazardly assembled combination of KiCad Git library
checkouts. To make everything work, some needed to be modified until
they loaded without errors on KiCad 4.0.x.

Since then, the KiCad team has released an official backport of the v5
libraries:
https://forum.kicad.info/t/kicad-4-0-x-backport-release-of-kicad-v5-library-is-now-available-on-github/10834

This change fixes paths and library names encoded in the project files
to reflect a KiCad 4.0.x installation configured to point at these
backported libraries.