Change U7 footprint to one that more closely matches the component
This involved adding a custom SOIC-28 footprint variant for the
AS6C6264.

This IC uses what it calls a "28-pin 330 mil SOP" package, which is not
in the standard libraries. Its closest match is KiCad's SOIC-28W, but it
is slightly wider and taller. This footprint and 3D model is based on
using the IC's datasheet and the reference datasheets from KiCad's
footprint and 3D model to determine small adjustments.

The replacement footprint is slightly wider, so some routing was
adjusted. However, it is mostly a drop-in replacement.

Fixes #1