Create new component to fix Q1/Q2/Q3 pin mapping issues
Apparently the S8050 included in the KiCad library was for the TO-92
version of the part, for which the pin mapping was correct. The part
we're using is a SOT-23, more completely called SS8050, and has
different specifications. So fixing the issue meant creating a new
library symbol with the right pin assignments.

The PCB layout has also been updated to reflect this change.

Fixes #5