Change SRAM address decoding to mirror it onto address $C000
To support the NES APU's DMC feature, we need sample memory placed
around address $C000. Since we're only using a tiny portion of our RAM,
the easiest way is to mirror the RAM onto both $0000 and $C000.

This can be accomplished with a spare NAND gate already in our circuit,
and the secondary non-inverting chip enable line of the SRAM chip. The
idea is that if either Y0 or Y6 is low, then the output of the NAND gate
will be high, and thus select the chip.

Fixes #7