Added TSL2591 component to the input board
The TSL2591 is an ambient light sensor, replacing the TSL250 originally
used on the main board. It is much better at detecting light through the
enclosure.

Added the component symbol, footprint, and approximate 3D model.
Integrated the component into the schematic.
Updated the PCB layout to include the component in the specific location
where the updated enclosure will have light holes.

The footprint should be accurate to the datasheet.
The 3D model is a generic DFN component in the same dimensions as the
TSL2591, but looks nothing like it.

Fixes #10