Initial NES and ESP code to handle I2C block data writes
This is an update to the protocol spec, and the code that
implements it.

The new capability being added is the ability to read and write blocks
of memory in the $C200-DFFF address range. For simplicity of
implmentation, data can be written up to 256 bytes at a time.
The purpose of this change is to support loading samples for the NES
APU's DMC feature.