NES APU DMC support for VGM playback
Adds support for the playback of sample data on the NES APU. This
includes loading data blocks from the VGM file into the 2A03's RAM, and
adjusting VGM commands that write to the address register ($4012) to
correctly point at that data.

This feature is one of the most complicated parts of VGM playback so
far, and thus will likely need refactoring for code clarity in the
future. It is essentially a virtual memory system that uses limited
static analysis.

The Nestronic cannot handle this part of the APU/VGM spec by simply
sending commands to the 2A03 verbatim, like the rest of it. This is
because the Nestronic does not have the same memory map as a real NES,
and because the VGM format allows the contents of that memory map to
change during playback.

This implementation begins by pre-scanning the complete VGM file for
data blocks and sample references. It copies the data blocks into a 32k
buffer representing the NES's $8000-$FFFF address space, keeping track
of the load time of individual segments. As references are captured,
these segments are collated into groups that are indexed by address and
load time. Once this scanning is complete, a "data state" structure is
made available to support the playback process.

Immediately prior to playback, these data groups are pre-loaded into the
available memory on the 2A03. If there is insufficient room to pre-load
any data, it is left to be dealt with later.

During the playback process, writes to the sample address register
($4012) are altered so that they point to the location that the
referenced data was actually loaded, instead of where it was when the
VGM was originally generated. In the event that a referenced data block
was not loaded, then a cost function is used to "evict" existing data
and then it is loaded incrementally during wait cycles.