Added RTC SRAM persistence support for the alarm set flag
Added functions to read/write to the MCP7940's SRAM region.

Added code to initialize the MCP7940 SRAM, using the ESP32's MAC address
as a flag to detect whether it has been initialized previously. This is
necessary because this region will likely contain random garbage on a
power-on reset.

Added code to store the alarm's "set" flag in this SRAM, so its
persistence is more closely tied to the clock time. This was also done
since this is a flag we expect a user to regularly fiddle with, and I'd
rather not wear the flash with this.