Initial NSF playback implementation
This implementation uses the Fake6502 emulator to "run" the code
contained within an NSF file, captures APU writes, and sends them to the
NES APU. It does not yet support DPCM, nor does it have any safety
against a malformed NSF file causing issues. Regular bankswitching is
supported, but FDS bankswitching is not.

It started as a proof-of-concept to see if this could be done, and will
be cleaned up and improved over time.