Add alarm trigger animation and fix touchpad input
Added a simple animation of the alarm clock icon, which is now shown
when the alarm is triggered. This provides visual indication of an alarm
in progress.

Also fixed touchpad input, which was probably broken as a side-effect of
the ESP-IDF 3.1 update. This input is used to trigger the snooze
function.