Updates to alarm triggering and snooze display
Changed the routine that detects an alarm trigger so that it's hopefully
not fooled by time skips. This works by tracking the current and
previous minute tick values, and comparing them against the alarm set
time.

Updated button handling so that an active alarm is only canceled by
actually pressing the alarm-off button.

Changed the snooze behavior so that it now shows a unique icon when an
active alarm is snoozed.