Try to prevent RTC time updates below the RTC output resolution
This change will attempt to avoid sending SNTP time updates to the RTC
when those updates change the existing value by less than approximately
one second. The goal of this is to reduce the likelyhood of sporatic
time changes without visible impact, which might get in the way of the
alarm trigger.