Workaround for missed RTC interrupt issue
Sometimes the RTC interrupt stops getting processed, which causes the
clock to stop updating by itself. The most likely cause is that an ISR
isn't triggered when the RTC MFP pin is low, which means the alarm flag
is never cleared, and the interrupt is not triggered again. While this
can happen due to startup timing issues, it has been observed to happen
at runtime. However, it is extremely rare and nearly impossible to
reproduce deliberately.

As an attempted workaround, the last time of the RTC interrupt event will
now be tracked. If more than 90 seconds passes without an event, the
GPIO polling task will inject it into the queue as if it had. Since this
interrupt normally happens every 60 seconds, and this issue is extremely
rare, this will hopefully clear it up.