Redesigned bankswitched NSF playback to cache unloaded banks
Complex NSF files were having issues with frame underruns, because the
time necessary to load new ROM blocks from the SD card exceeded the
available time budget of a single playback frame. This issue could
compound if blocks got twiddled between ROM banks during playback.

The first step to solving this issue was to separate loaded data blocks
from the address layout of the fake 6502. This was done by switching
from direct addressing of the memory data structure to indirect addressing
using a series of pointers corresponding to each of the 4k ROM banks.

The second, and arguably more complex, step was to allow allocating more
space for loaded ROM data than the 6502 banks can actually address and
to use an LRU cache to manage the use of that space. This means that
even if a particular NSF track tries to use more than 8 banks over the
course of its playback, it is probably not going to need to load a given
block of data from the SD card more than once.

The final step was to track the actual frame underrun time and make it
up by adjusting subsequent frame delays.